pyproject.toml
[project]
name = "media-upload"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.5",
    "django-dotenv==1.4.2",
    "django-softdelete>=0.11.5",
    "djangorestframework==3.16.1",
    "playwright==1.54.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv==1.1.1",
    "python-telegram-bot==22.3",
    "requests>=2.32.5",
    "uvicorn==0.35.0",
]
